Intro level update
-Added hazards: Magic orb turret + floor spikes
-Volume bug fixed
-Audio Manager enhanced
-Various music and sfx added
-Player healing added
-Enemy spawners animations

+++ Assets/Scripts/Character/Enemies/Enemy.cs
@@ -249,7 +249,7 @@ public class Enemy : Character
     }
 
     // Adds to base function to update healthbar
-    public override void SetHealth(int newValue)
+    public override void SetHealth(float newValue)
     {
         base.SetHealth(newValue);
         enemyHealthBar.UpdateBar(newValue, 0f, InitialHealth, true);
